# agent-world-lab/config/openapi/agent-world-lab.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgentWorldLab.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/chat/send": {
      "post": {
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSendResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSendResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSendResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/thread/{id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThreadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThreadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThreadResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ChatSendRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "threadId": {
            "type": "string",
            "nullable": true
          },
          "cloudProvider": {
            "type": "string",
            "nullable": true
          },
          "agentFrameworkMode": {
            "type": "string",
            "nullable": true
          },
          "modelBackendMode": {
            "type": "string",
            "nullable": true
          },
          "toolMode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChatSendResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "threadId": {
            "type": "string",
            "nullable": true
          },
          "reply": {
            "type": "string",
            "nullable": true
          },
          "traceEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceEventDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChatThreadResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "threadId": {
            "type": "string",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TraceEventDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}